refactor: Structure as a multi-page Streamlit app
Restructures the single-file application into a standard multi-page Streamlit application.

- The main script is renamed to `Home.py` and serves as the data upload page.
- A new `pages/Dashboard.py` script is created for the dashboard view.
- Helper functions for data processing and chart rendering are moved into a `utils/` directory to improve code modularity and maintainability.
