+++ utils/charting.py
@@ -1,51 +1,6 @@
 import streamlit as st
 import pandas as pd
-import json
 from streamlit_echarts import st_echarts
-import datetime
-
-# --- Page Configuration ---
-st.set_page_config(
-    page_title="Dynamic Dashboard Generator",
-    page_icon="📊",
-    layout="wide"
-)
-
-# --- Helper Functions ---
-
-def load_and_process_data(uploaded_file):
-    """
-    Loads data from the uploaded JSON file, flattens it, and processes it into a DataFrame.
-    """
-    if uploaded_file is None:
-        return pd.DataFrame()
-
-    try:
-        data = json.load(uploaded_file)
-        records = []
-        for entry in data:
-            flat_record = {
-                'pld': entry.get('pld'),
-                'asset_type': entry.get('asset_type'),
-                'timestamp': entry.get('timestamp')
-            }
-            if 'parameters' in entry and isinstance(entry['parameters'], dict):
-                flat_record.update(entry['parameters'])
-            records.append(flat_record)
-
-        df = pd.DataFrame(records)
-        if 'timestamp' not in df.columns:
-            st.error("The uploaded JSON must contain a 'timestamp' field in each record.")
-            return pd.DataFrame()
-
-        df['timestamp'] = pd.to_datetime(df['timestamp'], errors='coerce', utc=True)
-        df.dropna(subset=['timestamp'], inplace=True)
-        return df
-    except Exception as e:
-        st.error(f"An error occurred while processing the file: {e}")
-        return pd.DataFrame()
-
-# --- Chart Rendering Functions ---
 
 def render_line_chart(df, config, key):
     """Renders a line chart using ECharts."""
@@ -151,77 +106,3 @@ def configure_chart_dialog(df):
             st.session_state.charts.append(chart_config)
             st.session_state.show_modal = False
             st.rerun()
-
-# --- Main Application ---
-if 'data' not in st.session_state:
-    st.session_state.data = pd.DataFrame()
-if 'charts' not in st.session_state:
-    st.session_state.charts = []
-if 'show_modal' not in st.session_state:
-    st.session_state.show_modal = False
-
-st.sidebar.title("Navigation")
-menu_choice = st.sidebar.radio("Go to", ["Upload JSON", "Dashboard Preview"])
-
-if menu_choice == "Upload JSON":
-    st.title("📄 Upload Asset Data")
-    st.markdown("Upload a JSON file containing asset data to begin generating your dashboard.")
-    uploaded_file = st.file_uploader("Choose a JSON file", type="json")
-    if uploaded_file is not None:
-        with st.spinner('Processing data...'):
-            st.session_state.data = load_and_process_data(uploaded_file)
-            if not st.session_state.data.empty:
-                st.success("Data loaded successfully!")
-                st.dataframe(st.session_state.data.head())
-                st.write(f"Found **{len(st.session_state.data)}** records.")
-                st.session_state.charts = []
-
-elif menu_choice == "Dashboard Preview":
-    st.title("📊 Dashboard Preview")
-    df = st.session_state.data
-    if df.empty:
-        st.warning("Please upload a JSON file first using the 'Upload JSON' page.", icon="👈")
-    else:
-        st.header("Global Filters")
-        col1, col2, col3 = st.columns([2, 1, 1])
-        asset_types = df['asset_type'].unique().tolist()
-        global_selected_asset_type = col1.selectbox("Select Asset Type to View", ["All"] + asset_types)
-
-        min_date = df['timestamp'].dt.date.min()
-        max_date = df['timestamp'].dt.date.max()
-        from_date = col2.date_input("From date", min_date, min_value=min_date, max_value=max_date)
-        to_date = col3.date_input("To date", max_date, min_value=min_date, max_value=max_date)
-
-        df_timezone = df['timestamp'].dt.tz
-        from_datetime = pd.Timestamp(from_date, tz=df_timezone)
-        to_datetime = pd.Timestamp(to_date, tz=df_timezone) + pd.Timedelta(days=1) - pd.Timedelta(seconds=1)
-
-        st.markdown("---")
-
-        if st.button("➕ Add Chart", type="primary"):
-            st.session_state.show_modal = True
-
-        if st.session_state.show_modal:
-            configure_chart_dialog(df)
-
-        if not st.session_state.charts:
-            st.info("Your dashboard is empty. Click 'Add Chart' to get started.")
-        else:
-            cols = st.columns(2)
-            chart_index_to_render = 0
-            for i, config in enumerate(st.session_state.charts):
-                show_chart = (global_selected_asset_type == "All" or global_selected_asset_type == config['asset_type'])
-                if show_chart:
-                    chart_df = df[(df['asset_type'] == config['asset_type']) & (df['timestamp'] >= from_datetime) & (df['timestamp'] <= to_datetime)].copy()
-                    col_index = chart_index_to_render % 2
-                    with cols[col_index]:
-                        with st.container(border=True):
-                            chart_key = f"chart_{i}"
-                            if config['type'] == "Line Chart": render_line_chart(chart_df, config, chart_key)
-                            elif config['type'] == "Tabular Data": render_table(chart_df, config, chart_key)
-                            elif config['type'] == "Big Number": render_big_number(chart_df, config, chart_key)
-                            elif config['type'] == "Gauge": render_gauge(chart_df, config, chart_key)
-                            if st.button(f"Remove Chart {i+1}", key=f"remove_{i}"):
-                                st.session_state.charts.pop(i)
-                                st.rerun()
-                    chart_index_to_render += 1
